refactor: rs-dapi-client move dump code to dump.rs
