+++ packages/rs-dapi-client/src/dapi_client.rs
@@ -3,14 +3,10 @@
 use std::path::PathBuf;
 
 use backon::{ExponentialBuilder, Retryable};
-#[cfg(feature = "dump")]
-use hex::ToHex;
-#[cfg(feature = "dump")]
 use tonic::async_trait;
 use tracing::Instrument;
 
 use crate::{
-    mock::Key,
     transport::{TransportClient, TransportRequest},
     AddressList, CanRetry, RequestSettings,
 };
@@ -75,7 +71,7 @@ pub struct DapiClient {
     address_list: AddressList,
     settings: RequestSettings,
     #[cfg(feature = "dump")]
-    dump_dir: Option<PathBuf>,
+    pub(crate) dump_dir: Option<PathBuf>,
 }
 
 impl DapiClient {
@@ -86,60 +82,6 @@ impl DapiClient {
             settings,
             #[cfg(feature = "dump")]
             dump_dir: None,
-        }
-    }
-
-    #[cfg(feature = "dump")]
-    /// Prefix of dump files.
-    pub const DUMP_FILE_PREFIX: &str = "msg";
-
-    /// Define directory where dumps of all traffic will be saved.
-    ///
-    /// Each request and response pair will be saved to a JSON file in `dump_dir`.
-    /// Data is saved as [DumpData] structure.
-    /// Any errors are logged on `warn` level and ignored.
-    ///
-    /// Dump file name is generated by [DumpData::filename()].
-    ///
-    /// Useful for debugging and mocking.
-    /// See also [MockDapiClient::load()](crate::mock::MockDapiClient::load()).
-    #[cfg(feature = "dump")]
-    pub fn dump_dir(mut self, dump_dir: Option<PathBuf>) -> Self {
-        self.dump_dir = dump_dir;
-
-        self
-    }
-
-    /// Save dump of request and response to disk.
-    ///
-    /// Any errors are logged on `warn` level and ignored.
-    #[cfg(feature = "dump")]
-    fn dump_request_response<R: TransportRequest>(
-        request: R,
-        response: R::Response,
-        dump_dir: Option<PathBuf>,
-    ) where
-        R: serde::Serialize,
-        R::Response: serde::Serialize,
-    {
-        let path = match dump_dir {
-            Some(p) => p,
-            None => return,
-        };
-
-        let data = DumpData::new(request, response);
-
-        // Construct file name
-        // Path consists of current timestamp + hash of message
-        let filename = match data.filename() {
-            Ok(f) => f,
-            Err(e) => return tracing::warn!("unable to create dump file name: {}", e),
-        };
-
-        let file = path.join(filename);
-
-        if let Err(e) = data.save(&file) {
-            tracing::warn!("unable to write dump file {:?}: {}", path, e);
         }
     }
 }
@@ -212,101 +154,5 @@ impl Dapi for DapiClient {
         }
 
         result
-    }
-}
-
-#[cfg(feature = "dump")]
-#[derive(serde::Serialize, serde::Deserialize)]
-/// Data format of dumps created with [DapiClient::dump_dir].
-pub struct DumpData<T: TransportRequest> {
-    /// Fully qualified name of the type of request that was sent to DAPI.
-    pub request_type: String,
-    /// Request that was sent to DAPI.
-    pub request: T,
-    /// Response that was received from DAPI.
-    pub response: T::Response,
-}
-
-#[cfg(feature = "dump")]
-impl<T: TransportRequest> DumpData<T> {
-    /// Create new dump data.
-    ///
-    pub fn new(request: T, response: T::Response) -> Self {
-        let request_type = std::any::type_name::<T>().to_string();
-
-        Self {
-            request_type,
-            request,
-            response,
-        }
-    }
-    /// Return unique identifier (hex-encoded sha256) of the request.
-    /// Can be used to construct dump file name.
-    fn hash_hex(&self) -> Result<String, std::io::Error> {
-        Key::try_new(&self.request).map(|v| v.encode_hex())
-    }
-
-    /// Generate unique filename for this dump.
-    ///
-    /// Filename consists of:
-    ///
-    /// * [DapiClient::DUMP_FILE_PREFIX]
-    /// * current timestamp
-    /// * basename of the type of request, like `GetIdentityRequest`
-    /// * unique identifier (hash) of the request
-    pub fn filename(&self) -> Result<String, std::io::Error> {
-        let now = chrono::Utc::now().to_rfc3339_opts(chrono::SecondsFormat::Nanos, true);
-        let hash = self.hash_hex()?;
-
-        let request_type = self
-            .request_type
-            .split(':')
-            .last()
-            .unwrap_or(&self.request_type)
-            .replace('_', "-"); // remove underscore (separator used in file name)
-
-        let file = format!(
-            "{}_{}_{}_{}.json",
-            DapiClient::DUMP_FILE_PREFIX,
-            now,
-            request_type,
-            hash
-        );
-
-        Ok(file)
-    }
-
-    /// Load dump data from file.
-    pub fn load<P: AsRef<std::path::Path>>(file: P) -> Result<Self, std::io::Error>
-    where
-        T: for<'de> serde::Deserialize<'de>,
-        T::Response: for<'de> serde::Deserialize<'de>,
-    {
-        let f = std::fs::File::open(file)?;
-
-        let data: Self = serde_json::from_reader(f).map_err(|e| {
-            std::io::Error::new(
-                std::io::ErrorKind::InvalidData,
-                format!("unable to parse json: {}", e),
-            )
-        })?;
-
-        Ok(data)
-    }
-
-    /// Save dump data to file.
-    pub fn save(&self, file: &std::path::Path) -> Result<(), std::io::Error>
-    where
-        T: serde::Serialize,
-        T::Response: serde::Serialize,
-    {
-        let encoded = serde_json::to_vec_pretty(self).map_err(|e| {
-            std::io::Error::new(
-                std::io::ErrorKind::InvalidData,
-                format!("unable to serialize json: {}", e),
-            )
-        })?;
-
-        std::fs::write(file, encoded)
     }
 }
