+++ packages/rs-dapi-client/src/lib.rs
@@ -4,6 +4,8 @@
 
 mod address_list;
 mod dapi_client;
+#[cfg(feature = "dump")]
+pub mod dump;
 #[cfg(feature = "mocks")]
 pub mod mock;
 mod request_settings;
@@ -14,9 +16,9 @@ use futures::{future::BoxFuture, FutureExt};
 pub use http::Uri;
 
 pub use address_list::AddressList;
-#[cfg(feature = "dump")]
-pub use dapi_client::DumpData;
 pub use dapi_client::{DapiClient, DapiClientError};
+#[cfg(feature = "dump")]
+pub use dump::DumpData;
 pub use request_settings::RequestSettings;
 
 /// A DAPI request could be executed with an initialized [DapiClient].
